+++ src/main/java/gift/controller/ProductController.java
@@ -3,15 +3,9 @@ package gift.controller;
 import gift.dto.ProductRequestDto;
 import gift.dto.ProductResponseDto;
 import gift.dto.ProductStatusPatchRequestDto;
-import gift.exception.NotFoundByIdException;
-import gift.exception.RequestNotValidException;
 import gift.service.ProductService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,7 +16,6 @@ import java.util.List;
 public class ProductController {
 
     private final ProductService productService;
-    private static final Logger log = LoggerFactory.getLogger(ProductController.class);
 
     public ProductController(ProductService productService) {
         this.productService = productService;
@@ -38,12 +31,7 @@ public class ProductController {
     // For Admin
     @PostMapping("")
     public ResponseEntity<String> createProduct(
-            @Validated @RequestBody ProductRequestDto productRequestDto,
-            BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            throw new RequestNotValidException(bindingResult);
-        }
-
+            @Validated @RequestBody ProductRequestDto productRequestDto) {
         Long productId = productService.saveProduct(productRequestDto);
         return ResponseEntity.status(HttpStatus.CREATED).body("Successfully created id: " + productId);
     }
@@ -52,12 +40,8 @@ public class ProductController {
     @PutMapping("/{productId}")
     public ResponseEntity<String> updateProduct(
             @PathVariable Long productId,
-            @Validated @RequestBody ProductRequestDto productRequestDto,
-            BindingResult bindingResult
+            @Validated @RequestBody ProductRequestDto productRequestDto
     ) {
-        if (bindingResult.hasErrors()) {
-            throw new RequestNotValidException(bindingResult);
-        }
         productService.updateProduct(productId, productRequestDto);
         return ResponseEntity.ok().build();
     }
@@ -82,26 +66,5 @@ public class ProductController {
             ) {
         productService.updateProductStatus(productId, statusPatchRequestDto);
         return ResponseEntity.ok("Successfully Update Product's Status");
-    }
-
-    @ExceptionHandler(RequestNotValidException.class)
-    public ResponseEntity<String> handleRequestNotValidException(RequestNotValidException e) {
-        String message = e.getMessage();
-        log.trace(message);
-        return ResponseEntity.badRequest()
-                .body(message);
-    }
-
-    @ExceptionHandler(NotFoundByIdException.class)
-    public ResponseEntity<String> handleNotFoundByIdException(NotFoundByIdException e) {
-        log.trace(e.getMessage());
-        return ResponseEntity.status(HttpStatus.NO_CONTENT)
-                .body("Not Found by ID");
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<String> handleJsonParseError(HttpMessageNotReadableException e) {
-        log.trace(e.getMessage());
-        return ResponseEntity.badRequest().body("Invalid Request");
     }
 }
